Fix refactor after rebase
