+++ data_steward/constants/retraction/retract_utils.py
@@ -2,6 +2,8 @@ TABLE_ID = 'table_id'
 PERSON_ID = 'person_id'
 RESEARCH_ID = 'research_id'
 
+ALL_DATASETS = 'all_datasets'
+
 PID_QUERY = """
 (SELECT {pid}
 FROM `{pid_source}`)
